requirements.txt
opencv-python
numpy
scikit-image
scikit-learn
joblib
Pillow
reportlab
tk
